gtx: dfeEye() now returns an integer

+++ include/librorc/gtx.hh
@@ -94,7 +94,7 @@ namespace LIBRARY_NAME {
             bool isDomainReady();
             bool isLinkUp();
 
-            double dfeEye();
+            uint32_t dfeEye();
 
             void setTxDiffCtrl(uint32_t value);
             void setTxPreEmph( uint32_t value );
